Consider only positional messages for ts distribution
As previously mentioned, the s.d. of ts series is huge.  This is
probably in part because we are more likely to lose messages from
distant a/c than from nearer a/c, therefore we need to consider the
a/c's range when evaluating the quality of reception.

Therefore, this commit makes it so that only the intervals of positional
messages are considered.  G/s messages are still noted, but their
timestamps are not recorded because they'd add a second dimension to the
distribution of samples, making the distribution wider.

This commit does not yet consider range; I want to see how the ts
distribution changes when we consider only positional intervals first.

Bug fix: enable exclusion of old reports (as should have been the case)